[fix-02] - fixing masks from Membership
