+++ src/Pages/Public/MemberShip/index.jsx
@@ -15,6 +15,8 @@ import Modal from "../../../Components/Modal";
 import { useNavigate } from "react-router-dom";
 import { listOrgans } from "../../../Services/organService";
 import { isValidEmail } from "../../../Utils/validators";
+import { mascaraTelefone } from "../../../Utils/validators";
+import { mascaraCEP } from "../../../Utils/validators";
 
 const MemberShip = () => {
   const [email, setEmail] = useState("");
@@ -179,24 +181,6 @@ const MemberShip = () => {
   const mascaraRg = (rg) => {
     let formattedRG = rg.replace(/\D/g, "");
     return formattedRG;
-  };
-
-  const mascaraTelefone = (telefone) => {
-    let formattedTelefone = telefone.replace(/\D/g, "");
-    if (formattedTelefone.length > 11) {
-      formattedTelefone = formattedTelefone.slice(0, 11);
-    }
-    return formattedTelefone
-      .replace(/^(\d{2})(\d)/g, "($1) $2")
-      .replace(/(\d{4,5})(\d{4})$/, "$1-$2");
-  };
-
-  const mascaraCEP = (cep) => {
-    let formattedCEP = cep.replace(/\D/g, "");
-    if (formattedCEP.length > 8) {
-      formattedCEP = formattedCEP.slice(0, 8);
-    }
-    return formattedCEP.replace(/(\d{5})(\d)/, "$1-$2");
   };
 
   //eventos
